requirements.txt
certifi
cffi
charset-normalizer
cryptography
Django==3.2.5
djangorestframework==3.13.1
docopt==0.6.2
idna==3.3
# psycopg2
pycparser==2.21
PyMySQL==1.0.2
pytz==2021.3
# python==3.8.4
requests==2.27.1
sqlparse==0.4.1
typing-extensions==3.10.0.2
urllib3==1.26.8
wheel==0.37.1
wincertstore==0.2
yarg==0.1.9
django-cors-headers==3.10.1